+++ sitewit/constants.py
@@ -2,6 +2,11 @@ class CampaignServiceTypes(object):
     DIFM = 'difm'
 
 
+class BillingTypes(object):
+    AUTOMATIC = 'Automatic'
+    TRIGGERED = 'Triggered'
+
+
 CAMPAIGN_SERVICES = {
     CampaignServiceTypes.DIFM: 'DIFM Campaign'
 }